pyproject.toml
[tool.poetry]
name = "git-repo-sync"
version = "0.1.0"
description = ""
authors = ["Shaibu Shaibu <s.shaibu.jnr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
uvicorn = "^0.11.8"
gitpython = "^3.1.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.10.1"
pytest_mock = "^3.3.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
